% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educineqdata.R
\docType{data}
\name{data_country}
\alias{data_country}
\title{This dataset contains information about the available countries, their
corresponding country codes and the regions they belong to, which are used to
with \code{educineq} functions.}
\format{A data frame with 142 rows and 3 variables}
\usage{
data(data_country)
}
\description{
\itemize{
  \item country. Country name
  \item code. World Bank country code
  \item region. Macro-region to which the country belongs
}
}
\keyword{datasets}

