% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educineqdata.R
\docType{data}
\name{ineq_male25a}
\alias{ineq_male25a}
\title{This dataset contains some statistics about the distribution of educational
 attainments for male population aged over 25, taken from
 \url{www.educationdata.unican.es}. The data on population come from
 Barro and Lee database available at \url{http://www.barrolee.com}}
\format{A data frame with 1278 rows and 9 variables}
\usage{
data(ineq_male25a)
}
\description{
\itemize{
  \item country. Country name
  \item year
  \item code. World Bank country code
  \item region. Macro-region to which the country belongs
  \item mys. Mean years of schooling
  \item mld. Mean log deviation of education.
  \item theil. Theil index of education
  \item ge2. Generalized entropy measure of education
  \item pop. Total population, \url{http://www.barrolee.com}
}
}
\keyword{datasets}

